backend/requirements.txt
django
psycopg2-binary
djangorestframework
# google-generativeia